requirements.txt
adjustText==1.1.1
aeon==0.9.0
affine==2.4.0
aiohttp==3.9.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
args==0.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
basemap==1.4.0
basemap-data==1.3.2
basemap-data-hires==1.3.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
blosc2==2.7.0
cachetools==5.3.3
certifi==2024.7.4
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clint==0.5.1
cloudpickle==3.0.0
colorful==0.5.6
comm==0.2.2
contourpy==1.2.0
courlan==1.2.0
cycler==0.12.1
Cython==3.0.9
dask==2024.1.0
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dgl==2.1.0
distributed==2024.1.0
esig==0.9.8.3
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
fiona==1.9.6
Flask==3.0.3
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gensim==4.3.2
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
haversine==2.8.1
htmldate==1.8.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.1.6
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jusText==3.0.1
kiwisolver==1.4.5
libpysal==4.11.0
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.2
lxml_html_clean==0.1.1
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.3
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.10.1
numpy==1.26.3
opencensus==0.11.4
opencensus-context==0.1.3
overrides==7.4.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.4
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.18.0
pmdarima==2.0.4
prometheus-client==0.19.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
py-tgb==2.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyGSP==0.5.1
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pyunlocbox==0.5.2
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.1
QtPy==2.4.1
rasterio==1.3.10
ray==2.10.0
referencing==0.32.1
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scikit-base==0.6.2
scikit-learn==1.3.2
scikit-time==0.0.0
scipy==1.11.4
seaborn==0.13.1
seglearn==1.2.5
Send2Trash==1.8.2
shapely==2.0.4
signatory @ file:///signatory
six==1.16.0
sktime==0.25.0
smart-open==6.4.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12
tables==3.9.2
tblib==3.0.0
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tld==0.13
toolz==0.12.1
torch==2.1.2
torch-geometric-temporal==0.54.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch_cluster==1.6.3
torch_geometric==2.4.0
torchcde==0.2.5
torchdata==0.7.1
torchdiffeq==0.2.3
torchsde==0.2.6
tornado==6.4.1
tqdm==4.66.1
trafilatura==1.10.0
traitlets==5.14.3
trampoline==0.1.2
tsfresh==0.20.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
